+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/repositories/DeckRepository.java
@@ -1,0 +1,7 @@
+package ru.dpopkov.knowthenics.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import ru.dpopkov.knowthenics.model.Deck;
+
+public interface DeckRepository extends CrudRepository<Deck, Long> {
+}