+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/DeckService.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knowthenics.services;
+
+import ru.dpopkov.knowthenics.model.Deck;
+
+public interface DeckService extends CrudService<Deck, Long> {
+}