+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/FlashCardService.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knowthenics.services;
+
+import ru.dpopkov.knowthenics.model.FlashCard;
+
+public interface FlashCardService extends CrudService<FlashCard, Long> {
+}