+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/springdatajpa/FlashCardSDJpaService.java
@@ -1,0 +1,16 @@
+package ru.dpopkov.knowthenics.services.springdatajpa;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+import ru.dpopkov.knowthenics.model.FlashCard;
+import ru.dpopkov.knowthenics.repositories.FlashCardRepository;
+import ru.dpopkov.knowthenics.services.FlashCardService;
+
+@Service
+@Profile("spring-data-jpa")
+public class FlashCardSDJpaService extends AbstractSDJpaService<FlashCard> implements FlashCardService {
+
+    public FlashCardSDJpaService(FlashCardRepository flashCardRepository) {
+        super(flashCardRepository);
+    }
+}